requirements.txt
python-dotenv
cryptography
pandas
pyodbc
streamlit
streamlit-authenticator==0.2.2
msal
